requirements.txt
# Python package requirements for MIMIC-IV AF Study

# Core data science
pandas>=1.5.0
numpy>=1.23.0

# BigQuery
google-cloud-bigquery>=3.0.0
db-dtypes>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0

# Statistical analysis
scipy>=1.9.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.5.0

# Optional but recommended
openpyxl>=3.0.0  # For Excel export
